Performance and now respect the GIF frame delay specified
Removed support for running on a local PC just to get the performance where it is tolerable for showing each frame.

The frame delay is now being used to sleep between the frames being displayed.  Time is taken into account of processing the next frame while showing the previous frame.

Processing a full frame (i.e. all pixels have changed) and sending it to the OLED to be shown, takes ~185ms.  If you need to show GIFs that have a lot of changes, take this into account where the delay between frames might be longer than you expect.